Ui state change review.
Determining when the ui state change is called has moved to its own change listener listening to the underlying model and only for a change in queryText, a necessary change because the overlying model has been deliberately engineered not to fire if the queryText is the empty string and therefore not a valid thing to ask the server for results for.

The functions that used to be called "keyupAnimation" and "reverseAnimation" have been renamed to "expandedState" and "reducedState" respectively. The decision to reduce or expand the state has been moved to its own function out of the expand function.  The debounce on keyAnimation has been moved to where its used on the input event.

[rev. alex.scown]