Move user creation logic from NewUser to UsersPage
The reason for the refactor is that NewUser is a domain concept and hence should have no dependence on Selenium concepts.
These two classes are clearly quite tightly coupled to each other's implementations - is there a better way to do this?
Adding a type parameter (UsersPage<T extends NewUser>) is a possible option - UserService would also need to become generic, and possibly TestConfig.