[BIFHI-76] Refactored dates filtering so the dates filters state can be set from the saved search/snapshot models. [rev: daniel.grayling]
Previously, the QueryModel was used as the primary source of truth for the min and max dates, and the DatesFilterModel was updated from the QueryModel but also tracked the selected date range. Now, the primary source of truth for the user's date filter selections is the DatesFilterModel and the QueryModel is updated accordingly. This simplifies the control flow and allows us to set the dates filters from a saved search or snapshot.