Add dependency on mockui-hosted to POM for integration-tests-common
In order for hosted tests to run properly in IntelliJ, will need the hosted profile to be active in "Maven Projects" pane.
Dependency via profile so that integration-tests-common does not explicitly refer to hosted-only classes.
Having both onprem and hosted profiles active at the same time is allowed and encouraged when working within the IDE.