[BIFHI-118] Updated the related concepts view test to guard against a previous bug. [rev: none]
Previously, we were only listening to clicks on the <a> tag in cluster headings to filter by the concept. This caused a bug because it was possible to click on the <strong> tag beneath the <a>, resulting in "undefined" being used as a filter. This change adds a test to guard against this coming back.