[BIFHI-76] Clear the search bar when selecting a snapshot. [rev: alyona.ivanova]
This required refactoring the search page model into two models, one to contain the search bar input and related concepts, and another to contain the selected tab. This prevents a bug where the user clicked on a snapshot, which caused the search model selected tab to change, which caused the search model input text to be set to the empty string, which caused a new tab to be opened with a search for nothing.