[BIFHI-76] Fixed placeholder snapshot state token results view. [rev: none]
Scrolling after opening a snapshot tab caused exceptions to appear in the console. This is because a snapshot ServiceView instantiated a results view without appending it to the DOM, causing the scroll listener to throw exceptions.

Also moved HOD/IDOl specific overriding of the ResultsView constructor into the ResultsViewAugmentation View where it is instantiated.