[BIFHI-159] Updated server-side conversion of a SavedSearch to query and field text. [rev: will.byrne]
The same logic is duplicated in the client and server code. The SavedSnapshotController toQueryText method was missed when the equivalent client-side logic was updated. The toFieldText method was inconsistent with the client-side logic because multiple values for the same field were converted to "MATCH{value1}:field AND MATCH{value2}:field", not "MATCH{value1,value2}:field".

Also moved the methods to utility functions on the SavedSearch class and added unit tests.