[BIFHI-76] Enabled sunburst view for snapshots. [rev: jon.soul]
We need to send the snapshot state token with every parametric request. To facilitate this, the state tokens have been added to the query model.

Also, the parametric view is no longer responsible for fetching the parametric collection since the parametric collection is also used in the sunburst view.

Added tests for parametric views.