[CCUK-3647] Prevent fetching on scroll with small results sets
We now populate endOfResults after a search rather than before.
This means we don't need to get a second page before we can set
endOfResults.

Greatly simplify calculations for start and maxResults.
Setting maxResults to totalResults in the case of a small page feels
like a premature optimization. There was also an out by one error setting
start in one of the branches.

[for: alex.scown]