[BIFHI-243] Add reverse proxy support
The property server.reverseProxy enabled reverse proxy support
This opens an AJP connector on port 8009. The property server.ajp.port
configures this.
When using reverse proxy mode, the server will take the principal
from the reverse proxy and create a user in community with that
name. The will be given the FindUser role.

[rev: ivo.miller]