[BIFHI-265] Merged disabling of deleted databases with BIFHI.
In Find, the index filter for a deleted database is disabled (#146be3f). However, we can't use the query model to store the deleted databases in BIFHI as we did in Find since each tab has it's own query model. Instead, the relevant indexes collection model is given a deleted attribute.

[BIFHI-279]
[rev: alex.scown]