[BIFHI-281] Removed snapshot beans from the HOD context.
Snapshots will not work against HOD because it doesn't support state tokens, so we don't want to instantiate the snapshot beans.

Also moved most of the snapshot code into the idol package. Not all of it made sense to move since we share a database schema between HOD and IDOL.

Added an "abstract" SavedSearchRepository to make the AbstractSavedSearchService more generic.

[rev: jon.soul]