[FIND-27] Result renderer framework completed.
The result renderer now takes a "config" option which is provided in code by the results-view. This is the config file you would want to edit to provide a custom field for the case of a particular type of document model. Tests have been added and are intended to be a representative example of how you might edit the result-renderer-config to achieve this.

[rev: alex.scown]