[FIND-59] Remove search tabs
Remove the snapshot collection when the BI role is not
present.

We replace the savedQueryCollection with a regular
Backbone Collection and don't fetch it at startup.

The tab view is only created if the BI role is available. There's
also a styling shuffle so the containing div has no size if the view
is not present.

[rev: will.byrne]