[FIND-117] Show a minimum of 20 results in every sunburst with data  [rev alex.scown]
Includes now sending a null comparator to the sunburst's sort method to stop the segment data being sorted in descending order.
The order is now dependent only on the order of elements in the array. So can now use only a single padding segment.
Also changed the message for the padding segment to be more accurate.