[FIND-125] Pivot table columns
The table fixes the first column for easier scrolling. This
requires additional CSS because we're using Bootstrap.

We need to use visibility: hidden to make the measurements work.
This requires some div reordering to so messages appear in
the correct place.

We need a CSS bodge to fix an issue with HP Elements where
all li.active elements have transitions.

[rev: matthew.gordon]