[FIND-58] Make database optional
This is achieved by disabling Spring auto configuration
for the databases, and re-enabling it when the hp.find.enableBi
property is available. This in turn requires the disabling
of most of the beans which use the database.

The exclusions are configuration classes which are ConditionalOnClass
for Datasource, rather than ConditionalOnBean. Because we have these
classes on the classpath, the app blows up when we don't create a
Datasource.

We also don't grant the BI role to users if enableBI is false,
so the UI will hide the corresponding features.

[rev: ivo.miller]