[FIND-316] Fix tooltips
Tooltips were not removed when a parametric
value was selected. This was introduced in
c25ae2684819abefc788c160229fe19c7646610c but went
unnoticed as the parametric values did not change at that
time.

We need to reorder the statements in parametric-field-view
so that the element is still in the DOM when we destroy the
tooltip - removing first seems to delete the data required
by Bootstrap.

[rev: matthew.gordon]