[FIND-98] Time bar contents now controlled by clicking lhs widgets.
Data on time bar state (type = 'numeric' | 'date', unprettified field name) are now stored in timeBarModel. Each lhs widget has a change listener bound to timeBarModel.

On startup time bar is invisible. It is displayed on clicking a widget. It can be collapsed again by clicking the currently selected widget or the a close X icon in the upper-right corner of the time bar.