Don't pass null domains to client
The client no longer uses domains in IDOL mode,
and the nulls confuse it. Exclude null parameters
(name should never be null, and domain will only
be null when it is irrelevant) to fix issues in the client.

This should be revisited to use a more sensible server
side type.

[rev: ivo.miller]