Reverse-proxy mode needs both FindUser and FindBI roles
As part of the BIFHI 11.1 demo, I'm setting up a reverse-proxied instance of Find to show off the BIFHI features.  Defaulting to only the FindBI role is not enough to let users log in - they also need the FindUser role.

This branch should never go into `develop` or `master` - this is for the demo only.  A better solution will be required in the future.

[rev: none]