Change infrastructure to use Role annotations for on_prem
To use the role annotations, needed to subclass IdolFind and  IdolFindElementFactory (creating a subclass for each for BIFHI and for Find).
Create a second config with a non-BI user.
Test classes that use SavedSearchService need to typecast Application to BIIdolFind.
Changed Role, UserRole & UserRoleStrategy so that: 1) if a test has no annotation, can run with either role 2) if system variable not set, defaults to BIFHI.
Using the system variable is non-ideal but need this annotation v quickly - will later make it use config variable.