Add annotations & fix resulting problems
Filled in the test classes with annotations. Moved savedSearchService into IdolFind as abstract method and implemented in its child classes. This is needed when you use class level @Role, as typecasting is not safe then.  Added BOTH to UserRole so that, e.g. class can be annotated with BIFHI but one test can be BOTH. Moved IdolAutoCorrectITCase into bi package as it is BI-only.