Add concept input text to concepts instead of input text (FIND-584, subtask FIND-606)
We now add input text to the beginning of the concept array (we still clear the search as before for the non-bi user on change of text).

Removed all usages of query text model (in parallel with FIND-588 ticket updating database) except the overarching search model for the non-bi user's search box. In the future it may make sense to find a cleaner approach.

Removed routing to individual search terms for bi user as this does not make sense now that search text is no different from concepts. Routing currently still in place for non-bi user (full resolution of routing pending FIND-615).

Removed (at least temporarily) handling for updating search text in the top input box from inside a tab as there is no longer a definite use case for this, and it creates mess. This can be added back simply in some form if required.

We now surround concepts with brackets when generating query text instead of quotes as concepts entered via bar can contain operators etc. We consequently now surround clicked concepts with quotes.

[rev. matthew.gordon]