[FIND-215] Allow file upload
Uses dropzone.js to upload the files as this is what
Elements recommends.

We disable remove links because the default styling
hides them under the error tooltips.

The default strings for dropzone have been put
in the bundle so we control them.

By default the displayed error is just the server response,
so translate it into a bundle string.

[rev: will.byrne]