[FIND-578] Correct behaviour of concept clusters
Previous changes involved in [FIND-578] connected all concepts using
AND, whether they belonged to a cluster or not. This was undesirable,
because concepts which belong to clusters should be connected by spaces.

Example (before)
(Concept1) AND (Concept2a) AND (Concept2b) AND (Concept3)

Example (after this change)
(Concept1) AND (Concept2a Concept2b) AND (Concept3)

[rev: Alexander Scown]