SunburstITCase
New test for message appearing if no values on first search.
A more general search for testing the visibility of sunburst.
The hover in testHoveringOnGreySegmentGivesMessage was not hovering in the right place so used the hoveringOffSide method. Still relatively brittle.
Changed the parametric selectors tests are using to make it applicable to the test data. Having these hardcoded is not ideal.