Set to skip tests for Hosted
In FindSessionITCase, the assumption was not causing the test to be skipped on hosted (as it should).
Currently the roles are not set up for Hosted find.
To correctly use some of the tests in QueryTermITCase, the roles infrastructure needs to exist. Skip until fixed.
Further, the use of QueryTestHelpers is currently an impediment to fast test changes. Need to remove the dependence on this.