[FIND-402] Not allowing users to compare a selected search with itself.
Passed the original model cid of the selected model into the search-to-compare-view to compare it with the iterator model cid. The reason is because in compare-modal selected model is now being parsed and a new model gets created out of it with a new cid (which differs from the original, hence the modal view took it as a separate model).

[rev:will.byrne]