[FIND-671] Databases are now appearing on IE11.
IE11 has an issue of triggering an input event on input elements when placeholders are specified (and it set our filterModel's text to ''' and then trigger change events, etc). This was fixed by initialising the filterModel with '' even before the text input gets rendered, so when IE does trigger the input and tried to set the model value to '', no event will be triggered later on. Not so clever now, IE, are you?:)

[rev:matthew.gordon2]