[FIND-632] Topic Map now uses stateDontMatchIds
The topic map in the comparison view was not using the dont match ids for
the state tokens. These work by getting a token for each search, then a
token for exclusive to A and exclusive for B, the common to both
comparison uses match for A and B and passes in dont match for the
exlusives.

[rev Alex Scown]