Tests for Bugs FIND-686 and FIND-632
For testCommonToBothMap (FIND-632) needed to add methods to ResultsComparisonView to allow interaction with Topic Maps in this view. Also created enum AppearsInTopicMap to control switching between the maps for each and the 'Common To Both Map'.
Test also is having the same problem with the Concepts Panel form not actually submitting concept on submit on first try. Needs investigation and creating BIFindService is just a quick fix.
Also wrote testInlinePreviewClosesOnEdit for FIND-686.