Fixs for Table/TopicMap/DocumentPreview/IdolFilter/Results/SimilarDocumentsITCase [rev:matt.williamson]
TableITCase needsed more general searches and assumptions to ensure tests don't run if there aren't enough parametric values.
TopicMapITCase.testTopicMapShowTopicMap & DocumentPreviewITCase.testDetailedPreview needed to be more general.
IdolFilterITCase.testSearchForParametricFieldName was assuming that the name of 'goodField' filter category is not included in any other individual filters in different filter categories. For the badField, instead needs to get the first category whose filters do not container goodField. Also need to check that goodField name is present and badField is not present in any of the panel rather than assuming they will (not) be the 1st.
testNoMoreResultsFoundAtEnd was checking the visible number of results but should have been checking the total.
SimilarDocumentsITCase.testInfiniteScroll was waiting for the list view but should not have been. Also, limited the number of documents to 240, as could be thousands.