[FIND-697] Better handling of long field names
Some documents may contain long field names, which
can adversely affect document display. We can't just use
break-all as we're already using it for the field values
and browsers get confused, so we also need to add a
width to the column. This is lame, but unfortunately
there don't seem to be any alternatives.

Move the declarations to the less file
to reduce duplication in the template.

[rev: matthew.gordon]