testSingleFieldGivesCorrectTableValues & testTwoFieldsGiveCorrectTableValues
testSingleFieldGivesCorrectTableValues: to check the correct number of rows needed to assert that there were either as many rows as number of filters in current category.
testTwoFieldsGiveCorrectTableValues: need to find a 'good category' -> one that has between 1 and n filters & then check that are the correct number of columns with one already selected.
Also, updated getFilterNumber to return int not String.