[FIND-729, FIND-731] Set up the answered question result display
Created an answered question collection to store the question responses.
We will display one question answer pair above any promotions in the results list view if the query text returns any similar questions. For now we only display one, can increase to more in the future if desired.
Created a new list item template for the answered questions and a separate view to contain them.

[rev: matt.gordon]