Create OnPremSettingsPageITCase
Use settings tests and mockui from qa-infrastructure to include basic settings tests.
Renamed TopNavBar into NavBarSettings as it is now just the segment to the right of the navbar (dropdown and name). Search input present in nav bar for non-BI users is now not included.  Extended this with OnPremNavBarSettings, in which the dropdown has a link to the settings. Now get the searchInputBox directly.
Updated POM to get version 2.5.4 of qa-infrastructure.
Then use this to have set of tests in OnPremSettingsPageITCase

[rev:matt.williamson]