[FIND-795] Grid system for Dashboards
We have implemented a system that uses absolute positions, when the
display: grid feature is supported on all browsers we support we should
look into changing to that as it is designed for our exact purpose.

I have implemented a basic staticContent widget for testing pusposes and
we currently treat all widgets as one of those for the time being.

rev Alex Scown