[FIND-804] allow access to dashboard saved search
We have added a new request mapping in the idol saved search controller to
allow the retrieval of a single saved search. This new method checks
agains the ids in the dashboard config file to ensure it is listed there
as a valid id.

We consume this in the widget abstract view and set the saved search
model as an attribute of the model.

rev Alex Scown