[FIND-798] Add click through for widgets
When a widget is clicked the user will be routed to the backing saved search /
snapshot. We changed the config file slightly to add the type for the
saved search and this required a minor change to the controllers to filter
out the apropriate widgets in the validation of the ids.

rev Alex Scown