[FIND-932] export pagination reprints headers
The export has been altered to now paginate in the export service and the
writeHeaders boolean flag has now been altered into a method and in the
CSV case we have added the byte mark there. This is called once before we
start paginating over the results.

rev Matt Gordon