Restore meta-filtering of parametric fields (FIND-979)
Added a filtering collection for parametric values
Whenever the meta-filter is changed, we send a value restricted parametric value request off to the server for (non-numeric/date) parametric fields whose names don't match the meta-filter. When this is done we update the value filtering collection with the combination of these value restricted results and the unchanged results for the matched fields.
We re-run meta-filtering upon changes to the parametric collection.

[rev. matthew.gordon]