[FIND-894] Added fullscreen to dashboards
The dashboards now use the html5 fullscreen api to enter "noc" mode.
This is implemented with minimal selenium testing as selenium cannot
handle fullscreen yet. When the spec is finalised this will need to be
looked at to remove the experimental prefixes and tested in cases the
specification has changed.

rev Matt Gordon