[FIND-917] Data fetch error handling for widgets
[rev: Matthew Gordon]

Saved search widgets now monitor their fetch promises for failures and
display an error message in lieu of the visualizer if the initialisation
or update fail at any stage. If an exception has been sent to the client,
the error message includes it.

Added exhaustive Jasmine tests for the initialization/update sequence. The
suite monitors the widget's behaviour over three cycles and checks all
combinations of success, failure, and a fetch being interrupted by another
scheduled update (happens if the fetch takes a long time or the update
interval is short).