Improve initial loading behaviour of widgets (FIND-990)
We now display a loading spinner until widget initialisation is complete.
We don't wait till the data finishes loading, as we have the existing updating loading spinner for that.
We trigger it early for the map, immediately before the map tile is rendered, as otherwise the map is rendered wrong.
This does not fix loading handling for widgets when fetching data.

[rev. matthew.gordon]