[FIND-905] Dashboad click through
We have fixed click through behaviour for the dashboard widgets. When a
user who doesnt own the search clicks, they are taken to a read only tab
which displays the search criteria in the same way as the snapshot tab,
but doesnt show the rename button and the delete button is replaced with a
close button.

rev Matt Gordon