[FIND-1053] Debounce database selection at click level
Previously we were debouncing the update of the selected
databases collection. This meant there was a delay of 500ms between
populating the collection for the first time and being able to search.
By debouncing the clicks, this problem goes away.

We still need to update the UI on click, which means updating all the
category checkboxes, so we use much of the existing functionality but
against a different array, then reconcile the changes after the debounce
period.

Should be resilient to external changes through the updateSelectedDatabases
method.

[rev: matthew.gordon]