[FIND-1008] Topic Map fixed height
The topic map was set to a fixed height. this was problematic for
responsive reasons and broke the topic map widget. We have now switched it
to have an extra class for dashboards which sets the height to 100% and on
the main results view we have set it to be 100vh - the total height of the
bars above it.

rev Matt Gordon