[FIND-719] Fixed "No more results" message appearing after auto-correct.
This was happening because the results view was listening to
any change event on the query model for its fetch function.
When the search comes back with an auto-correct, the spellcheck
view sets an attribute on the query model, triggering a change
event. The response to this unnecessary query caused "No more
results found" to be displayed.

[rev: ivo.miller]