[FIND-1058] Do fewer updates
Restructure find-search and results-view-container so that they will
either render or update, not both. This works because in the initial
case we call update from elsewhere, so it's pointless work.

The topic map now needs to call update from render again to make this
work. Other views seem to require no changes.

Should work with any view as the first view.

[rev: matthew.gordon]