Replace failing Se test; unit test entity topic map
[rev: Matthew Gordon]

A Selenium test was failing intermittently because of results being
data-dependent.

In the process I discovered that we were not correctly using the
maxResults configuration in the entity topic map view, leading to
configured values being outright ignored in some circumstances. This
has been corrected and is covered by tests.