When a user drags the full data range off the chart in one go it causes errors (FIND-1222)
Add an invisible rectangle to the chart which covers the whole graph area.
Add drag and zoom behaviour to this rectangle instead of the chart.
Ensure the rectangle is always inserted above any group elements in the dom so that line interaction always works correctly.
Prevent immediate redraw if the user drags past the last point on the graph (subsequent redraw after data fetch will still occur).

[rev. rosie.taylor]