[FIND-945] Prevent comparison tab hover covering back button.
This was happening because the button has a box shadow. The
tab mouseover was covering this shadow. Adding a margin to the
button fixes this.

Also improved layout of the map comparison view. This included
adding a "magic number" height which could potentially be
fixed by a flex column layout in a later ticket.

[rev: alex.scown]