[FIND-969] Move where current user is added to the database
Instead of having the user added in the login success handler which wouldn't be called if we were using a reverse proxy, added a login listener that creates a user on application event.
Also fixed the tests to add a user to the database as part of the database set up for the test.

[rev: alex.scown]