[FIND-969] Run flyway migration before running integration tests
Extract the post update function to a file the will only be run if the isBiEnabled application property is set to true.

For the tests, set the migration on update to do nothing and create a flyway migration strategy bean that runs the migration from Spring Boot like it used to. The database will never have any data that needs migrating for the tests, so the fact that we can't query Community from here doesn't matter, but now the database is correctly set up for the tests.

 [rev: alex.scown]