[FIND-969] Change how we add current user to database in test setup
@DatabaseSetup was being called before @Sql, so the main test data was being added before we had added the current user to the database leading to errors.

The @Sql now will either add or update and does not error if the same user has been added already. We can then add the user with user_id of 1 back to the main database setup data.

This also fixes a bug in core where without the integration test setup, the shared-to-user.xml was missing the first user.

[rev: alex.scown]