Ensure all foreign keys have ON DELETE CASCADE in database
The cascade was not being added to the search_parametric_ranges foreign key, causing the migration to fail if any data present in that table.

Have also added to shared_to_users foreign keys so that if in future we want to rely on the cascade delete it is implemented for all the tables.

[rev: matt.gordon]