[FIND-1595] ControlPoint: make credentials optional
It turns out that the new CP APIs created for use by Find don't require any
authentication (though all pre-existing APIs do).  'The securityinfo is the
credentials.'  So the user shouldn't have to configure Find with any
credentials - leaving the credentials system in place though, now that it's
implemented anyway.